Gate the Turnkey test cases behind a turnkey feature
The Turnkey signer-backend cases skipped at runtime when the TURNKEY_* env
vars were absent, silently reporting ok and requiring per-test skip plumbing.
Gate them on a breez-itest turnkey feature instead: default builds do not
contain the cases at all (6 tests), and building with --features turnkey
compiles them (13 tests) and requires the credentials, failing loudly when
they are unset. Enabling the feature is the explicit opt-in, so no env-based
skip logic remains.

The Turnkey harness (per-test wallet provisioning, cleanup guard, reaper, SDK
builder) moves to its own module gated once at the declaration, the itest
crate only enables the SDK's turnkey features through its own feature, and
SdkInstance's guard slot uses an uninhabited placeholder type without the
feature so construction sites need no per-field gating.

Note: a CI job running the Turnkey cases must pass --features turnkey.
